♻️ (signer-solana): Move device-action factories to internal, re-export from api
Keep the public factory API stable while restoring clean layering: the factory
implementations now live in internal/app-binder/device-action/ and the api files
just re-export them. SolanaAppBinder imports the internal factories directly, so
there is no more runtime internal -> api dependency (which created a module
cycle). Tests moved next to the implementations.
