+++ packages/signer/signer-solana/src/internal/app-binder/device-action/SignMessageDeviceActionFactory.test.ts
